+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # HEAD
 
+- (**Breaking change**) Turn `rank` and `comm` parameters into
+  keywords for `generate_noise_mpi`
+- (**Breaking change**) Use `nothing` instead of `missing` for MPI
+  communicators
+- Ensure that `generate_noise_mpi` produces different seeds for each
+  polarimeter
+- (**Breaking change**) Make `destripe` return a `DestripingResults`
+  structure instead of a tuple
 - Remove the dependency on `Quaternions.jl` [#38](https://github.com/lspestrip/Stripeline.jl/pull/38)
 - Update the instrument database with the number of the polarizer for each horn [#37](https://github.com/lspestrip/Stripeline.jl/pull/37)
 - Make the 1/f slope always in the range 0..2 [#34](https://github.com/lspestrip/Stripeline.jl/pull/34)